refactor nav bar
