+++ src/components/auth/AuthProgressBar.js
